requirements.txt
praw
kafka-python
psycopg2-binary
python-dotenv
apache-airflow
nltk
